requirements.txt
flask>=2.0.1
python-dotenv>=0.19.0
google-generativeai>=0.3.0
torch>=1.10.0
torchaudio>=0.10.0
transformers>=4.16.0
librosa>=0.8.1
soundfile>=0.10.3
pydub>=0.25.1
matplotlib>=3.4.0
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
Werkzeug>=2.0.0